0.4.2a
A display bug is fixed with a method I am not 100% confident with.
The series of games that the movie poster that I put into a comment in the code is a little offensive to computer geeks with a stereotypical "computer geek" character, but I have played a little of two of the games in the series and I haven't seen him.  In fact, I just happen to know about the games and have even played some of them thanks to someone I know who loves them.